Move ml_uncertainty into examples.ML_EXPERIMENTS because this belongs together
